ui: move cover art metadata controls into dedicated panel
